requirements-colab.txt
numpy
opencv-python
pillow
matplotlib
scikit-image
scipy
shapely
svgwrite
cairosvg
tqdm
networkx
pyyaml
requests
openai
transformers
accelerate
bitsandbytes
einops
sentencepiece